Suggesting this micro optimization:

instead of using a string key concatenation, it is possible to use a weakmap and an empty object as key.

Note that we can't use the `routeOptions` itself because it is being destroyed by fastify at:

https://github.com/fastify/fastify/blob/97df2dfbd9411c3a68fc2d8107f13cf32f5f797f/lib/route.js#L306-L310

---

The `seed` reference is a unique object per route, so we can access it later. The benchmark `autocannon -p 3 -d 30 -c 10 http://localhost:3000/hello` says:

## main

```
┌─────────┬──────┬──────┬───────┬──────┬─────────┬─────────┬───────┐
│ Stat    │ 2.5% │ 50%  │ 97.5% │ 99%  │ Avg     │ Stdev   │ Max   │
├─────────┼──────┼──────┼───────┼──────┼─────────┼─────────┼───────┤
│ Latency │ 0 ms │ 0 ms │ 1 ms  │ 1 ms │ 0.04 ms │ 0.29 ms │ 32 ms │
└─────────┴──────┴──────┴───────┴──────┴─────────┴─────────┴───────┘
┌───────────┬─────────┬─────────┬─────────┬────────┬───────────┬──────────┬─────────┐
│ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%  │ Avg       │ Stdev    │ Min     │
├───────────┼─────────┼─────────┼─────────┼────────┼───────────┼──────────┼─────────┤
│ Req/Sec   │ 46,975  │ 46,975  │ 52,575  │ 53,343 │ 51,595.74 │ 1,951.46 │ 46,958  │
├───────────┼─────────┼─────────┼─────────┼────────┼───────────┼──────────┼─────────┤
│ Bytes/Sec │ 8.83 MB │ 8.83 MB │ 9.89 MB │ 10 MB  │ 9.7 MB    │ 367 kB   │ 8.83 MB │
└───────────┴─────────┴─────────┴─────────┴────────┴───────────┴──────────┴─────────┘

Req/Bytes counts sampled once per second.
# of samples: 30

1548k requests in 30.01s, 291 MB read
```

## this branch

```
┌─────────┬──────┬──────┬───────┬──────┬─────────┬─────────┬───────┐
│ Stat    │ 2.5% │ 50%  │ 97.5% │ 99%  │ Avg     │ Stdev   │ Max   │
├─────────┼──────┼──────┼───────┼──────┼─────────┼─────────┼───────┤
│ Latency │ 0 ms │ 0 ms │ 0 ms  │ 1 ms │ 0.03 ms │ 0.16 ms │ 11 ms │
└─────────┴──────┴──────┴───────┴──────┴─────────┴─────────┴───────┘
┌───────────┬─────────┬─────────┬─────────┬─────────┬─────────┬──────────┬─────────┐
│ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%   │ Avg     │ Stdev    │ Min     │
├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼──────────┼─────────┤
│ Req/Sec   │ 59,391  │ 59,391  │ 64,447  │ 66,303  │ 64,184  │ 1,663.34 │ 59,389  │
├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼──────────┼─────────┤
│ Bytes/Sec │ 11.2 MB │ 11.2 MB │ 12.1 MB │ 12.5 MB │ 12.1 MB │ 312 kB   │ 11.2 MB │
└───────────┴─────────┴─────────┴─────────┴─────────┴─────────┴──────────┴─────────┘

Req/Bytes counts sampled once per second.
# of samples: 30

1926k requests in 30.01s, 362 MB read
```

The server:

```js
import Fastify from 'fastify'
import zeroTrust from './dist/index.js'

const fastify = Fastify()

// Register the plugin
await fastify.register(zeroTrust)

// Public route
fastify.get('/hello', {
  allowIf: async (request) => true
}, async (request, reply) => {
  return { hello: 'world' }
})

await fastify.listen({ port: 3000 })
```

## Notes:

- The `excludedRoutes` option could be evaluated during the `onRoute` execution to gain even more speed
- `as any` is my YOLO out of my working hours